backend/requirements.txt
google-adk==1.21.0
playwright==1.51.0
litellm==1.80.11
opik==1.9.96
rich
pydantic-settings
termcolor
fastapi[standard]